requirements.txt
yt-dlp==2026.2.21
faster-whisper==1.2.1
google-genai==1.65.0
opencv-python==4.13.0.92